requirements.txt
asgiref==3.5.2
decouple==0.0.7
Django==3.2.15
django-crispy-forms==1.14.0
django-environ==0.9.0
djangorestframework==3.13.1
Pillow==9.2.0
python-decouple==3.6
pytz==2022.1
sqlparse==0.4.2
